test(tab-menu): Add example tests and move into file within component
