+++ packages/pancake-uikit/package.json
@@ -46,6 +46,7 @@
     "styled-components": "^5.2.0"
   },
   "dependencies": {
+    "@testing-library/user-event": "^13.0.0",
     "@types/lodash": "^4.14.162",
     "@types/styled-system": "^5.1.10",
     "lodash": "^4.17.20",
